+++ CHANGELOG.md
@@ -36,12 +36,14 @@ This version add the test for the next enspoints:
 ### Added
 - enpoint for get an order by id in `src/api/v1/apps/orders/router.py`
 - enpoint for get all orders in `src/api/v1/apps/orders/router.py`
+- Tests:
+  - get and post orders endpoints in `tests/integration/test_order_post_get_endpoints.py`
 
 ### Changed
 - N/A
 
 ### Fixed
-- N/A
+- Added the user_id to the current_user payload for fixture auth user in `tests/conftest.py`
 
 ## [1.10.0] 2026-06-05
 This version add the orders module